[refactor] UserInfo->UserProfile 변경으로 인해 변경된 코드 변 (#399)
