+++ src/main/java/org/sopt/app/presentation/user/UserResponseMapper.java
@@ -1,14 +1,11 @@
 package org.sopt.app.presentation.user;
 
 import java.util.List;
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
+import org.mapstruct.*;
 import org.sopt.app.application.playground.dto.PlaygroundProfileInfo;
 import org.sopt.app.application.operation.OperationInfo;
 import org.sopt.app.application.app_service.AppServiceInfo;
-import org.sopt.app.application.user.UserInfo;
-import org.sopt.app.presentation.user.UserResponse.AppService;
+import org.sopt.app.presentation.user.UserResponse.*;
 
 @Mapper(
         componentModel = "spring",
@@ -20,9 +17,9 @@ public interface UserResponseMapper {
     UserResponse.MainView ofMainView(PlaygroundProfileInfo.MainView user, OperationInfo.MainView operation,
             Boolean isAllConfirm);
 
-    UserResponse.Nickname of(UserInfo.Nickname nickname);
+    UserResponse.Nickname of(Nickname nickname);
 
-    UserResponse.ProfileMessage of(UserInfo.ProfileMessage profileMessage);
+    UserResponse.ProfileMessage of(ProfileMessage profileMessage);
 
     UserResponse.Generation ofGeneration(PlaygroundProfileInfo.UserActiveInfo userActiveInfo);
